refactor(cli): isolate remote tailscale install helpers
